Streamify/backend/requirements.txt
Django>=5.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
